// src/app/Models/registroEstudiante.ts
export class registroEstudiante{
    NombreEst: string;
    ApellidoEst: string;
    FechaNacimientoEst:Date;
    TipoDocumentoEst:string;
    NumeroDocumentoEst:number;
    CelularEst: number;
    CorreoEst: string;
    DireccionEst:string;
    PasswordEst: string;
    estado: string;
    NombreUsuarioEst: string;
}